[Frontend] Move Order and Repositories into Advanced dialog
Extracts OrderTab and RepositoriesTab from AdminPlugins into standalone
AdvancedOrderSection and AdvancedRepositoriesSection components, wraps
them in a gear-button-triggered AdvancedPluginsDialog, and wires a
?advanced= search param so Task 14 can deep-link into either section.
